Modulo pokemon y demas

+++ src/app/tasks/tareas/tareas.component.ts
@@ -1,7 +1,9 @@
-import { Component, OnInit, DoCheck } from '@angular/core';
+
+import { Component, OnInit, DoCheck, ViewChild } from '@angular/core';
 import { TareaService } from '../tarea.service';
 import { Task } from '../tarea';
 import { Router } from '@angular/router';
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 
 
 @Component({
@@ -13,9 +15,15 @@ export class TareasComponent implements OnInit, DoCheck {
   public totalTasks: number;
   public tasksList: Array<Task>;
   public order: boolean;
+  public selectedTask: Task;
+  private closeResult;
+  @ViewChild('contentModal') contentmodal: HTMLElement;
 
-  constructor(private tareaService: TareaService, private router: Router) {
+  constructor(private tareaService: TareaService,
+              private router: Router,
+              public modalWindow: NgbModal) {
     this.order = false;
+    this.selectedTask = null;
   }
 
 
@@ -29,8 +37,16 @@ export class TareasComponent implements OnInit, DoCheck {
   }
 
   deleteTask(taskToDelete: Task) {
-    this.tareaService.deleteTask(taskToDelete);
-    this.tasksList = [...this.tareaService.tasks];
+    this.modalWindow.open(this.contentmodal, {ariaLabelledBy: 'modal-basic-title'}).result.then((result) => {
+      this.closeResult = result;
+      if (!this.closeResult) {
+        return null;
+        } else {
+          this.tareaService.deleteTask(taskToDelete);
+          this.tasksList = [...this.tareaService.tasks];
+          this.selectedTask = null;
+        }
+    }, () => { });
   }
 
   goToForm() {
